Print next prime number by specifying the position
This Python program accepts 2 values (only integer values) as input. The 1st value is the current number and the 2nd value is the position of the next prime number that you need. For example if the input is (1,2) the program accepts 1 as current value and returns the prime number residing in the 2nd position next to number 1.