Search within a dictionary with a Search String
This Python Program Accepts a Search String as Input. A Sample Dictionary is created and a Input search string is given to search for the value within the dictionary.

When a matching value is found, the function returns the (Name) key of that corresponding item.

This program does a case insensitive search.